fix: wrap board_to_tensor in torch.FloatTensor before .to() in _calibrate_elo
board_to_tensor() returns a numpy array, which has no .to() method.
The bug crashes the training loop every CALIBRATION_INTERVAL (50)
steps when _calibrate_elo() runs.

+++ training/trainer.py
@@ -316,7 +316,7 @@ class Trainer:
                     if not legal:
                         break
                     with torch.no_grad():
-                        tensor = board_to_tensor(board).to(self.device)
+                        tensor = torch.FloatTensor(board_to_tensor(board)).to(self.device)
                         pred_policy, _ = self.model(tensor.unsqueeze(0))
                     scores = {}
                     for m in legal:
